A program to Flattening the linked list
In this program we used a main node which is connected to next and down pointers which is representing the next and the branch pointer. The program works recursively to merge the subnodes using merge function which is similar to merge sort. The flatten method here merges subnodes from right to left reccursively and merge method to merge the sorted parts.

Time complexity:- O(n log n)